tried to optimize a lot of the bigint stuff
inlined a lot of assembly and took out some +0 and -0 opperations that otherwise had use but not in this case and were auto generated by the derive feature. we'll see if this improves performance at all even though it's still stuck behind a reference. Unfortunately I'm not sure if it's easy or possible with Jolt as a target to remove the trait bound of BigInt from the MontgomeryConfig